+++ exporters/trace/zipkin/model.go
@@ -146,7 +146,10 @@ var extraZipkinTags = []string{
 }
 
 func toZipkinTags(data *export.SpanSnapshot) map[string]string {
-	m := make(map[string]string, len(data.Attributes)+len(extraZipkinTags))
+	m := make(map[string]string, len(data.Attributes)+len(data.Resource.Attributes())+len(extraZipkinTags))
+	for _, kv := range data.Resource.Attributes() {
+		m[(string)(kv.Key)] = kv.Value.Emit()
+	}
 	for _, kv := range data.Attributes {
 		m[(string)(kv.Key)] = kv.Value.Emit()
 	}